Features related to SymbolParsing. Improved Decoding interface and Scanning on Memory Snapshots.

+++ Process/Platform/Windows/WinapiLink.hpp
@@ -52,6 +52,9 @@ public:
 
 	bool attach(uint32_t procID) override
 	{
+		if (isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		HANDLE tmpHandle = OpenProcess(PROCESS_ALL_ACCESS, false, procID);
 
 		if (tmpHandle == INVALID_HANDLE_VALUE)
@@ -79,6 +82,9 @@ public:
 
 	bool isAlive(bool& isAliveState) const override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		uint32_t exitCode;
 		if (this->getExitCode(exitCode))
 		{
@@ -90,6 +96,9 @@ public:
 
 	bool terminate()
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		DWORD exitCode = 1;
 		if (TerminateProcess(this->_hProcess, exitCode))
 		{
@@ -103,6 +112,9 @@ public:
 
 	bool suspend()
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		typedef NTSTATUS( NTAPI* fnNtSuspendProcess )(HANDLE ProcessHandle);
 		static fnNtSuspendProcess suspendProcess = nullptr;
 
@@ -127,6 +139,9 @@ public:
 
 	bool resume()
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		typedef NTSTATUS( NTAPI* fnNtResumeProcess )(HANDLE ProcessHandle);
 		static fnNtResumeProcess resumeProcess = nullptr;
 
@@ -150,7 +165,10 @@ public:
 	}
 
 	bool getExitCode(uint32_t& exitCode) const override
-	{ 
+	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		DWORD eCode;
 		if (GetExitCodeProcess(this->_hProcess, &eCode))
 		{
@@ -164,17 +182,42 @@ public:
 
 	bool getMainProcessImage(ProcessImage& processImage) const override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		return setError(retrieveProcessMainImage(this->_procID, processImage));
 	}
 
 	bool getProcessImage(const std::string& name, ProcessImage& processImage) const override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		return setError(retrieveProcessImage(this->_procID, name, processImage));
 	}
 
 	bool getAllProcessImages(std::vector<ProcessImage>& processImages) const override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		return setError(retrieveAllProcessImages(this->_procID, processImages));
+	}
+
+	bool getExportSymbols(const ProcessImage& image, std::vector<ImageSymbol>& symbols) const override
+	{
+		if (!isAttached())
+			return setError(EPlatformError::InvalidState);
+
+		return setError(retrieveExportSymbols(image, symbols));
+	}
+
+	bool getImportSymbols(const ProcessImage& image, std::vector<ImageSymbol>& symbols) const override
+	{
+		if (!isAttached())
+			return setError(EPlatformError::InvalidState);
+
+		return setError(retrieveImportSymbols(image, symbols));
 	}
 
 	//=== Process Query ===//
@@ -230,6 +273,9 @@ public:
 
 	bool read(uint64_t addr, size_t size, void* buffer) const override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		size_t bytesRead;
 		if (ReadProcessMemory(this->_hProcess, (LPVOID)addr, buffer, size, (SIZE_T*)&bytesRead) && bytesRead != 0)
 			return true;
@@ -239,6 +285,9 @@ public:
 
 	bool write(uint64_t addr, size_t size, const void* buffer) override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		size_t bytesRead;
 		if (WriteProcessMemory(this->_hProcess, (LPVOID)addr, buffer, size, (SIZE_T*)&bytesRead) && bytesRead != 0)
 			return true;
@@ -248,6 +297,9 @@ public:
 
 	bool queryMemory(uint64_t addr, MemoryRegion& memoryRegion) const override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		MEMORY_BASIC_INFORMATION mbi{};
 		if (VirtualQueryEx(this->_hProcess, (LPVOID)addr, &mbi, sizeof(mbi)) != 0)
 		{
@@ -259,6 +311,9 @@ public:
 
 	bool virtualProtect(uint64_t addr, size_t size, EMemoryProtection newProtect, EMemoryProtection* oldProtect) override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		DWORD winNewProtect = toWinProtect(newProtect);
 		DWORD winOldProtection = 0;
 		if (VirtualProtectEx(this->_hProcess, (LPVOID)addr, size, winNewProtect, (PDWORD)(&winOldProtection)))
@@ -272,6 +327,9 @@ public:
 
 	uint64_t virtualAllocate(uint64_t addr, size_t size, EMemoryProtection protection) override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		auto winProt = toWinProtect(protection);
 		addr = (uint64_t)VirtualAllocEx(this->_hProcess, NULL, size, MEM_COMMIT | MEM_RESERVE, winProt);
 		
@@ -283,6 +341,9 @@ public:
 
 	bool virtualFree(uint64_t addr) override
 	{
+		if (!isAttached())
+            return setError(EPlatformError::InvalidState);
+
 		if (VirtualFreeEx(this->_hProcess, (LPVOID)addr, 0, MEM_RELEASE))
 		{
 			return true;
